pmem: handle requests for aligned memory correctly
The calculation of next_bit is incorrect, which can return
pointers to blocks which are not actually aligned as specified.
Additionally, there is no guarantee that i = 0 (bit 0) will be
aligned on the first iteration. Correct the calculation of
next_bit to give the correct block and actually start on the
start_bit calculated in the previous function.

Change-Id: I5c23de8d3095cd20cf66a789bae641c70ef45cb4
CRs-Fixed: 288000
Signed-off-by: Laura Abbott <lauraa@codeaurora.org>