msm: acpuclock: Fix acpuclk_wait_for_irq()
On msm7x2x, the acpuclk_wait_for_irq() API is supposed to ensure
that the CPU is running <= 128MHz per hardware requirements.
Previously, this 128MHz value was hardcoded, so it didn't work
for frequency tables that didn't have 128MHz rows.

Fix this by finding the highest frequency row <= 128MHz during
init and selecting it when the API is called and the current
frequency is not already <= 128MHz.

Change-Id: I91848b7ab231b5def27ada40870da643d0e296b5
Signed-off-by: Matt Wagantall <mattw@codeaurora.org>

Conflicts:

	arch/arm/mach-msm/acpuclock.c